[refactor] ValuedEnum string -> enum 변환 로직 공통화 - #1
