+++ novelservice/src/main/java/com/iucyh/novelservice/common/enumtype/valuedenum/ValuedEnum.java
@@ -1,17 +1,6 @@
-package com.iucyh.novelservice.common.enumtype;
+package com.iucyh.novelservice.common.enumtype.valuedenum;
 
 public interface ValuedEnum {
 
     String getValue();
-
-    default <T extends Enum<T> & ValuedEnum> T fromValue(String value, Class<T> enumClass) {
-        if (value == null || value.isBlank()) return null;
-
-        for (T e : enumClass.getEnumConstants()) {
-            if (e.getValue().equals(value)) {
-                return e;
-            }
-        }
-        return null;
-    }
 }
